created views folder and moved the component directly into it.
