+++ src/views/ArkPullsCalculator.vue
